requirements.txt
streamlit==1.34.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.4.2
folium==0.16.0
google-generativeai==0.5.4
openpyxl
